add: timeline editor with hooks to reduce code size for TimelineEditor
