+++ src/components/video-editor/timeline/model/timelineModel.ts
@@ -3,8 +3,6 @@ import type {
 	AnnotationRegion,
 	AudioRegion,
 	ClipRegion,
-	SpeedRegion,
-	TrimRegion,
 	ZoomMode,
 	ZoomRegion,
 } from "../../types";
